Wrap logging setup with a function and split info/debug/warning/error
