+++ prototype/sky/registry.py
@@ -3,8 +3,8 @@ import collections
 
 import sky
 from sky import clouds
-import logging
-logger = logging.getLogger(__name__)
+from sky.logging import init_logger
+logger = init_logger(__name__)
 
 _CLOUDS = [
     clouds.AWS(),
@@ -20,7 +20,6 @@ def fill_in_launchable_resources(task: sky.Task):
             launchable[resources] = [resources]
             continue
         if resources.cloud is not None:
-            logger.info(resources.cloud)
             launchable[
                 resources] = resources.cloud.get_feasible_launchable_resources(
                     resources)
